requirements.txt
fastapi
uvicorn
pydantic
httpx

# Groq - LLM for query refinement
groq

# Local embeddings (no API needed)
sentence-transformers

# Vector DB
faiss-cpu
numpy

# Data ingestion
pandas

# Retry logic
tenacity

# Environment variable loading
python-dotenv
